bip/sip example + linear updates + parse_args migration (#17)
* bip vs sip

* move parse_args to helpers

* parse args inherited, missing example added

* argparse inherited

* adding inverse run

* missing imports

* clean up
