+++ setup.cfg
@@ -64,6 +64,7 @@ pub =
 console_scripts =
     mud_examples = mud_examples.runner:run
     generate_poisson_data = mud_examples.poisson:run
+    mud_run_inv = mud_examples.runner:run_inv
     mud_run_lin = mud_examples.runner:run_lin
     mud_run_pde = mud_examples.runner:run_pde
     mud_run_ode = mud_examples.runner:run_ode
