+++ src/mud_examples/plotting.py
@@ -4,7 +4,6 @@
 import logging
 
 from matplotlib import pyplot as plt
-from matplotlib import cm
 import matplotlib.gridspec as gridspec
 plt.rcParams['figure.figsize'] = 10,10
 plt.rcParams['font.size'] = 16
@@ -14,10 +13,6 @@ plt.rcParams['font.size'] = 16
 # matplotlib.backend = 'Agg'
 
 import numpy as np
-from mud.util import null_space
-from mud.funs import mud_sol, map_sol
-from mud.norm import full_functional, norm_input, norm_data, norm_predicted
-from mud.plot import make_2d_unit_mesh
 from mud_examples.helpers import check_dir
 
 _logger = logging.getLogger(__name__) # TODO: make use of this instead of print
@@ -35,188 +30,6 @@ def fit_log_linear_regression(input_values, output_values):
     regression_line = 10**(slope * x + intercept)
     return regression_line, slope
 
-
-def plot_2d_contour_example(A=np.array([[1, 1]]), b=np.zeros([1, 1]),  # noqa: C901
-                            cov_11=0.5, cov_01=-0.25,
-                            initial_mean=np.array([0.25, 0.25]),
-                            alpha=1, omega=1, obs_std=1,
-                            show_full=True, show_data=True,
-                            show_est=False, param_ref=None, compare=False,
-                            fsize=42, figname='latest_figure.png', save=False,
-                            ):
-    """
-    alpha: float in [0, 1], weight of Tikhonov regularization
-    omega: float in [0, 1], weight of Modified regularization
-    """
-    # mesh for plotting
-    N, r = 250, 1
-    X, Y, XX = make_2d_unit_mesh(N, r)
-    inputs = XX
-
-    std_of_data = [obs_std]
-    obs_cov = np.diag(std_of_data)
-    observed_data_mean = np.array([[1]])
-    initial_cov = np.array([[1, cov_01], [cov_01, cov_11]])
-
-    assert np.all(np.linalg.eigvals(initial_cov) > 0)
-
-    z = full_functional(A, XX, b, initial_mean, initial_cov, observed_data_mean, observed_cov=obs_cov)
-    zp = norm_predicted(A, XX, initial_mean, initial_cov)
-    zi = norm_input(XX, initial_mean, initial_cov)
-    zd = norm_data(A, XX, b, observed_data_mean, observed_cov=obs_cov)
-    # sanity check that all arguments passed correctly:
-    assert np.linalg.norm(z - zi - zd + zp) < 1E-8
-
-    # plotting contours
-    z = (alpha * zi + zd - omega * zp)
-    mud_a = np.argmin(z)
-    map_a = np.argmin(alpha * zi + zd)
-
-    # get mud/map points from minimal values on mesh
-    mud_pt = inputs[mud_a, :]
-    map_pt = inputs[map_a, :]
-
-    msize = 500
-    ls = (np.linalg.pinv(A) @ observed_data_mean.T).ravel()
-    if show_data:
-        plt.contour(inputs[:, 0].reshape(N, N),
-                    inputs[:, 1].reshape(N, N),
-                    (zd).reshape(N, N), 25,
-                    cmap=cm.viridis, alpha=0.5, vmin=0, vmax=4)
-        plt.axis('equal')
-
-        s = np.linspace(-2 * r, 2 * r, 10)
-
-        if A.shape[0] < A.shape[1]:
-            # nullspace through least-squares
-            null_line = null_space(A) * s + ls.reshape(-1, 1)
-            plt.plot(null_line[0, :], null_line[1, :],
-                     label='Solution Contour',
-                     lw=2, color='xkcd:red')
-            if not show_full:
-                plt.annotate('Solution Contour', (0.1, 0.9),
-                             fontsize=fsize, backgroundcolor="w")
-
-    if show_full:
-        plt.contour(inputs[:, 0].reshape(N, N),
-                    inputs[:, 1].reshape(N, N),
-                    z.reshape(N, N), 50,
-                    cmap=cm.viridis, alpha=1.0)
-    elif alpha + omega > 0:
-        plt.contour(inputs[:, 0].reshape(N, N),
-                    inputs[:, 1].reshape(N, N),
-                    (alpha * zi - omega * zp).reshape(N, N), 100,
-                    cmap=cm.viridis, alpha=0.25)
-    plt.axis('equal')
-
-
-    if alpha + omega > 0:
-        plt.scatter(initial_mean[0], initial_mean[1],
-                    label='Initial Mean',
-                    color='k', s=msize)
-        if not show_full:
-            plt.annotate('Initial Mean',
-                         (initial_mean[0] + 0.001 * fsize, initial_mean[1] - 0.001 * fsize),
-                         fontsize=fsize, backgroundcolor="w")
-        else:
-            if param_ref is not None:
-                plt.scatter(param_ref[0], param_ref[1],
-                    label='$\\lambda^\\dagger$',
-                    color='k', s=msize, marker='*')
-            if compare:
-                plt.annotate('Truth',
-                         (param_ref[0] + 0.0005 * fsize, param_ref[1] + 0.0005 * fsize),
-                         fontsize=fsize, backgroundcolor="w")
-
-        show_mud = omega > 0 or compare
-
-        if show_full:
-            # scatter and line from origin to least squares
-            plt.scatter(ls[0], ls[1],
-                        label='Least Squares',
-                        color='xkcd:blue',
-                        marker='d', s=msize, zorder=10)
-            plt.plot([0, ls[0]], [0, ls[1]],
-                     color='xkcd:blue',
-                     marker='d', lw=1, zorder=10)
-            plt.annotate('Least Squares',
-                         (ls[0] - 0.001 * fsize, ls[1] + 0.001 * fsize),
-                         fontsize=fsize, backgroundcolor="w")
-
-            if show_est:  # numerical solutions
-                if omega > 0:
-                    plt.scatter(mud_pt[0], mud_pt[1],
-                                label='min: Tk - Un', color='xkcd:sky blue',
-                                marker='o', s=3 * msize, zorder=10)
-                if (alpha > 0 and omega != 1):
-                    plt.scatter(map_pt[0], map_pt[1],
-                                label='min: Tk', color='xkcd:blue',
-                                marker='o', s=3 * msize, zorder=10)
-
-            if (alpha > 0 and omega != 1):  # analytical MAP point
-                map_pt_eq = map_sol(A, b, observed_data_mean,
-                                    initial_mean, initial_cov,
-                                    data_cov=obs_cov, w=alpha)
-                plt.scatter(map_pt_eq[0], map_pt_eq[1],
-                            label='MAP', color='xkcd:orange',
-                            marker='x', s=msize, lw=10, zorder=10)
-
-                if compare:  # second map point has half the regularization strength
-                    plt.annotate('MAP$_{\\alpha}$',
-                                 (map_pt_eq[0] - 0.004 * fsize, map_pt_eq[1] - 0.002 * fsize),
-                                 fontsize=fsize, backgroundcolor="w")
-
-                else:
-                    plt.annotate('MAP$_{\\alpha}$',
-                                 (map_pt_eq[0] + 0.0001 * fsize, map_pt_eq[1] - 0.002 * fsize),
-                                 fontsize=fsize, backgroundcolor="w")
-
-            if show_mud:  # analytical MUD point
-                mud_pt_eq = mud_sol(A, b, observed_data_mean,
-                                    initial_mean, initial_cov)
-                plt.scatter(mud_pt_eq[0], mud_pt_eq[1],
-                            label='MUD', color='xkcd:brown',
-                            marker='*', s=2 * msize, lw=5, zorder=10)
-                plt.annotate('MUD',
-                             (mud_pt_eq[0] + 0.001 * fsize, mud_pt_eq[1] - 0.001 * fsize),
-                             fontsize=fsize, backgroundcolor="w")
-
-        if A.shape[0] < A.shape[1]:
-            # want orthogonal nullspace, function gives one that is already normalized
-            v = null_space(A @ initial_cov)
-            v = v[::-1]  # in 2D, we can just swap entries and put a negative sign in front of one
-            v[0] = - v[0]
-
-            if show_full and show_mud:
-                # grid search to find upper/lower bounds of line being drawn.
-                # importance is the direction, image is nicer with a proper origin/termination
-                s = np.linspace(-1, 1, 1000)
-                new_line = (v.reshape(-1, 1) * s) + initial_mean.reshape(-1, 1)
-                mx = np.argmin(np.linalg.norm(new_line - initial_mean.reshape(-1, 1), axis=0))
-                mn = np.argmin(np.linalg.norm(new_line - mud_pt_eq.reshape(-1, 1), axis=0))
-                plt.plot(new_line[0, mn:mx], new_line[1, mn:mx], lw=1, label='projection line', c='k')
-        elif show_full:
-            plt.plot([initial_mean[0], ls[0]],
-                     [initial_mean[1], ls[1]],
-                     lw=1, label='Projection Line', c='k')
-
-    #     print(p)
-
-    plt.axis('square')
-    plt.axis([0, r, 0, r])
-#     plt.legend(fontsize=fsize)
-    plt.xticks(fontsize=0.75 * fsize)
-    plt.yticks(fontsize=0.75 * fsize)
-    plt.tight_layout()
-    if save:
-        if '/' in figname:
-            fdir = ''.join(figname.split('/')[:-1])
-            check_dir(fdir)
-        plt.savefig(figname, dpi=300)
-        plt.close()
-    else:
-        plt.show()
-#     plt.title('Predicted Covariance: {}'.format((A@initial_cov@A.T).ravel() ))
 
 
 def plot_decay_solution(solutions, model_generator, sigma, prefix,
@@ -287,9 +100,7 @@ def plot_decay_solution(solutions, model_generator, sigma, prefix,
             fdir = filename.split('/')[0]
             check_dir(fdir)
             plt.savefig(filename, bbox_inches='tight')
-            plt.close()
-        else:
-            plt.show()
+        # plt.show()
 
 
 def plot_experiment_equipment(tolerances, res, prefix, fsize=32, linewidth=5,
@@ -457,3 +268,35 @@ def plot_scalar_poisson_summary(res, measurements, prefix, lam_true, fsize=32, s
             plt.savefig(f'{_prefix}_reference_solution.png', bbox_inches='tight')
         else:
             plt.show()
+
+
+def plotChain(mud_chain, ref_param, color='k', s=100):
+    num_steps = len(mud_chain)
+    current_point = mud_chain[0]
+    plt.scatter(current_point[0], current_point[1], c='b', s=s)
+    for i in range(0, num_steps):
+        next_point = mud_chain[i]
+        points = np.hstack([current_point, next_point])
+        plt.plot(points[0, :], points[1, :], c=color)
+        current_point = next_point
+
+    plt.ylim([0, 1])
+    plt.xlim([0, 1])
+#     plt.axis('off')
+    plt.scatter(ref_param[0], ref_param[1], c='r', s=s)
+
+
+def plot_contours(A, ref_param, subset=None,
+                  color='k', ls=':', lw=1, fs=20, w=1, s=100, **kwds):
+    if subset is None:
+        subset = np.arange(A.shape[0])
+    A = A[np.array(subset), :]
+    numQoI = A.shape[0]
+    AA = np.hstack([null_space(A[i, :].reshape(1, -1)) for i in range(numQoI)]).T
+    for i, contour in enumerate(subset):
+        xloc = [ref_param[0] - w * AA[i, 0], ref_param[1] + w * AA[i, 0]]
+        yloc = [ref_param[0] - w * AA[i, 1], ref_param[1] + w * AA[i, 1]]
+        plt.plot(xloc, yloc, c=color, ls=ls, lw=lw, **kwds)
+        plt.annotate('%d' % (contour + 1), (xloc[0], yloc[0]), fontsize=fs)
+
+
